+++ src/components/ids-module-nav/ids-module-nav-display-mode-mixin.ts
@@ -1,10 +1,7 @@
 import { attributes } from '../../core/ids-attributes';
-import { DISPLAY_MODE_TYPES } from './ids-module-nav-common';
 
 import type { IdsModuleNavDisplayMode } from './ids-module-nav-common';
 import type { IdsConstructor } from '../../core/ids-element';
-
-import { stripTags } from '../../utils/ids-xss-utils/ids-xss-utils';
 
 export interface DisplayModeHandler {
   onDisplayModeChange?(currentValue: string | false, newValue: string | false): void;
@@ -45,17 +42,15 @@ const IdsModuleNavDisplayModeMixin = <T extends Constraints>(superclass: T) => c
    * @param {IdsModuleNavDisplayMode | null} value Display Mode setting
    */
   set displayMode(value: IdsModuleNavDisplayMode | null) {
-    let safeValue: any = null;
-    if (typeof value === 'string') {
-      safeValue = stripTags(value, '');
+    let safeValue: IdsModuleNavDisplayMode = false;
+    if (typeof value === 'string' && (value === 'expanded' || value === 'collapsed')) {
+      safeValue = value;
     }
 
     const currentValue = this.displayMode;
     if (currentValue !== safeValue) {
-      if (DISPLAY_MODE_TYPES.includes(safeValue)) {
-        if (safeValue !== false) {
-          this.setAttribute(attributes.DISPLAY_MODE, `${safeValue}`);
-        }
+      if (safeValue !== false) {
+        this.setAttribute(attributes.DISPLAY_MODE, `${safeValue}`);
       } else {
         this.removeAttribute(attributes.DISPLAY_MODE);
         safeValue = false;