+++ src/components/ids-module-nav/ids-module-nav-settings.ts
@@ -6,7 +6,6 @@ import IdsModuleNavTextDisplayMixin from './ids-module-nav-text-display-mixin';
 import IdsMenuButton from '../ids-menu-button/ids-menu-button';
 import IdsMenuItem from '../ids-menu/ids-menu-item';
 
-import type IdsLocale from '../ids-locale/ids-locale';
 import type IdsPopupMenu from '../ids-popup-menu/ids-popup-menu';
 import type IdsText from '../ids-text/ids-text';
 
@@ -22,6 +21,8 @@ const Base = IdsModuleNavTextDisplayMixin(
  * IDS Module Nav Settings Component
  * @type {IdsModuleNavSettings}
  * @inherits IdsModuleNavItem
+ * @mixes IdsModuleNavDisplayModeMixin
+ * @mixes IdsModuleNavTextDisplayMixin
  * @part expander - this accoridon header's expander button element
  * @part header - the accordion header's root element
  * @part icon - the accordion header's icon element
@@ -74,17 +75,16 @@ export default class IdsModuleNavSettings extends Base {
     }
 
     this.setPopupmenuType();
+    this.setPopupmenuCoords();
 
     this.menuEl.popup.arrow = 'none';
-    this.menuEl.popup.x = 8;
-    this.menuEl.popup.y = 12;
 
-    if (this.displayMode === 'collapsed') {
-      this.menuEl.popup.align = 'right, bottom';
-    }
-    if (this.displayMode === 'expanded') {
-      this.menuEl.popup.align = 'top, left';
-    }
+    this.menuEl.popup.onPlace = (popupRect: DOMRect) => {
+      if (this.displayMode === 'collapsed') {
+        popupRect.y -= 8;
+      }
+      return popupRect;
+    };
   }
 
   /**
@@ -113,6 +113,21 @@ export default class IdsModuleNavSettings extends Base {
     });
   }
 
+  setPopupmenuCoords() {
+    const isRTL = this.menuEl.popup.localeAPI.isRTL();
+
+    this.menuEl.popup.x = 8;
+    this.menuEl.popup.y = 12;
+    this.menuEl.popup.useRight = isRTL;
+
+    if (this.displayMode === 'collapsed') {
+      this.menuEl.popup.align = 'right, bottom';
+    }
+    if (this.displayMode === 'expanded') {
+      this.menuEl.popup.align = `${isRTL ? 'bottom' : 'top'}, ${isRTL ? 'right' : 'left'}`;
+    }
+  }
+
   /**
    * @param {string | undefined | null} variantName name of the new colorVariant
    */
@@ -134,10 +149,8 @@ export default class IdsModuleNavSettings extends Base {
     if (this.textNode) this.textNode.audible = val !== 'default';
   }
 
-  onLanguageChange = (locale?: IdsLocale | undefined) => {
+  onLanguageChange = () => {
     if (!this.menuEl || !this.menuEl.popup) return;
-    if (this.displayMode === 'collapsed') {
-      this.menuEl.popup.align = locale?.isRTL() ? 'left' : 'right';
-    }
+    this.setPopupmenuCoords();
   };
 }