+++ doc/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### 1.0.0-beta.17 Fixes
 
 - `[Calendar]` Add `disableSettings` property to calendar. ([#1471](https://github.com/infor-design/enterprise-wc/issues/1471))
+- `[Modal]` Add `showCloseButton` setting to modal. ([#1527](https://github.com/infor-design/enterprise-wc/issues/1527))
 
 ## 1.0.0-beta.16